NAV-22906: Flytter utbetalingsperiode type til egen fil
