+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/FulltidBarnehageplassAugust2024Alert.test.tsx
@@ -7,11 +7,8 @@ import { kjønnType } from '@navikt/familie-typer';
 import { FulltidBarnehageplassAugust2024Alert } from './FulltidBarnehageplassAugust2024Alert';
 import { type IGrunnlagPerson, PersonType } from '../../../typer/person';
 import { Målform } from '../../../typer/søknad';
-import {
-    type IUtbetalingsperiodeDetalj,
-    type Utbetalingsperiode,
-    Vedtaksperiodetype,
-} from '../../../typer/vedtaksperiode';
+import type { Utbetalingsperiode } from '../../../typer/utbetalingsperiode';
+import { type IUtbetalingsperiodeDetalj, Vedtaksperiodetype } from '../../../typer/vedtaksperiode';
 
 describe('FulltidBarnehageplassAugust2024Alert', () => {
     test('skal ikke rendre komponent om ingen utbetalingsperioder har tom i august 2024', () => {
