+++ src/frontend/komponenter/Fagsak/Behandlingsresultat/Oppsummeringsboks.tsx
@@ -18,7 +18,7 @@ import type {
     IRestValutakurs,
 } from '../../../typer/eøsPerioder';
 import { EøsPeriodeStatus, KompetanseResultat } from '../../../typer/eøsPerioder';
-import type { Utbetalingsperiode } from '../../../typer/vedtaksperiode';
+import type { Utbetalingsperiode } from '../../../typer/utbetalingsperiode';
 import { dateTilFormatertString, Datoformat } from '../../../utils/dato';
 import {
     formaterBeløp,
