+++ src/frontend/typer/behandling.ts
@@ -15,12 +15,12 @@ import type {
 } from './tilbakekrevingsbehandling';
 import type { ITotrinnskontroll } from './totrinnskontroll';
 import type { IRestEndretUtbetalingAndel } from './utbetalingAndel';
+import type { Utbetalingsperiode } from './utbetalingsperiode';
 import type {
     IRestKorrigertEtterbetaling,
     IRestKorrigertVedtak,
     IVedtakForBehandling,
 } from './vedtak';
-import type { Utbetalingsperiode } from './vedtaksperiode';
 import type { IRestPersonResultat, IRestStegTilstand } from './vilkår';
 import type { IRestBrevmottaker } from '../komponenter/Fagsak/Personlinje/Behandlingsmeny/LeggTilEllerFjernBrevmottakere/useBrevmottakerSkjema';
 import type { IsoDatoString } from '../utils/dato';
