+++ src/frontend/typer/fagsak.ts
@@ -2,7 +2,7 @@ import type { BehandlingÅrsak } from './behandling';
 import type { BehandlingKategori } from './behandlingstema';
 import type { INøkkelPar } from './common';
 import type { ITilbakekrevingsbehandling } from './tilbakekrevingsbehandling';
-import type { Utbetalingsperiode } from './vedtaksperiode';
+import type { Utbetalingsperiode } from './utbetalingsperiode';
 import type { VisningBehandling } from '../komponenter/Fagsak/Saksoversikt/visningBehandling';
 
 // Enum
